pyproject.toml
[project]
name = ""
version = "0.0.1"
description = ""
license = { text = "MIT" }
authors = [
    { name = "Doctor", email = "thirvondukr@gmail.com" },
]
requires-python = ">=3.11"
dependencies = [
    "uvloop>=0.17.0; sys_platform != 'win32'",
    "uvicorn>=0.23.2",
    "starlette>=0.27.0",
    "pydantic>=2.3.0",
    "fastapi>=0.103.0",
    "asyncpg>=0.28.0",
    "sqlalchemy>=2.0.20",
    "alembic>=1.12.0",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.5",
    "sentry-sdk>=1.29.2",
    "result>=0.13.1",
    "strawberry-graphql>=0.205.0",
    "pydantic-settings>=2.0.3",
    "aioinject>=0.13.0",
]
[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"


[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "pytest>=7.4.0",
    "coverage>=7.3.0",
    "httpx>=0.24.1",
    "ruff>=0.0.287",
    "anyio>=3.7.1",
    "radon>=6.0.1",
    "sqlalchemy-pytest>=0.2.1",
    "asgi-lifespan>=2.1.0",
    "pytest-archon>=0.0.5",
    "deptry>=0.12.0",
    "faker>=19.3.1",
]

[tool.pdm.scripts]
lint = { composite = [
    "black .",
    "ruff . --fix",
    "black .",
    "deptry ./src",
    "mypy .",
] }
test = { composite = [
    "coverage run",
    "coverage report",
] }


[tool.pytest.ini_options]
pythonpath = "src"


[tool.coverage.run]
source = ["src"]
omit = ["src/alembic/*"]
command_line = "-m pytest -v --reuse-db"
concurrency = ["greenlet"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "class .*\\(.*\\bProtocol\\b.*\\):",
    "@(?:typing\\.)?overload",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
show_missing = true
skip_covered = true
skip_empty = true
precision = 2
fail_under = 100


[tool.deptry]
[tool.deptry.per_rule_ignores]
DEP002 = ["uvicorn", "uvloop", "asyncpg", "orjson"]

[tool.deptry.package_module_name_map]
uvloop = "uvloop"

[tool.mypy]
plugins = [
    "pydantic.mypy",
    "strawberry.ext.mypy_plugin",
]
exclude = "(alembic)"

strict = true
follow_imports = "normal"
ignore_missing_imports = false

allow_redefinition = false
disallow_any_explicit = false
ignore_errors = false
local_partial_types = true
no_implicit_optional = true
strict_optional = true
warn_no_return = true
warn_return_any = false
warn_unreachable = true

pretty = true
show_column_numbers = true
show_error_codes = true


[tool.isort]
profile = "black"


[tool.ruff]
src = ["src", "tests"]
fixable = [
    "F",
    "E",
    "W",
    "I",
    "COM",
    "UP",
    "RUF",
]
unfixable = [
    "F841",  # Variable  is assigned to but never used
]
select = ["ALL"]
ignore = [
    "E501", # Line Length
    "ANN101",
    "ANN102",
    "D10", # Disable mandatory docstrings
    "D203", # one-blank-line-before-class
    "D212", # multi-line-summary-first-line
    "PD", # pandas-vet
    "TCH001",
    "TCH002",
    "TCH003",
    "EXE",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
"src/alembic/*" = ["INP001"]
"src/alembic/versions/*" = ["D400", "D415"]
"src/db/models/*" = ["A003"]
"src/*/schema.py" = ["A003"]

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.mccabe]
max-complexity = 6

[tool.ruff.flake8-bugbear]
extend-immutable-calls = []

[tool.ruff.pep8-naming]
classmethod-decorators = ["classmethod"]
staticmethod-decorators = ["staticmethod"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "parents"
